server/my_project/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.5.0
djangorestframework==3.15.2
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
